% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clone_f_selection_sweep}
\alias{clone_f_selection_sweep}
\title{Clone frequencies of a sweeping clone at various time points}
\format{
A data frame with 50 rows and 7 variables:
\describe{
  \item{reactions}{Total number of reactions at present time.}
  \item{time}{Gillespie time.}
  \item{cells}{Total number of cells.}
  \item{clone1}{Number of cells in the ancestral clone.}
  \item{clone2}{Number of cells in the subclone.}
  \item{seed}{Random seed used to generate the simulation.}
  \item{start_time}{Time point (reactions) at which the subclone was added.}
}
}
\usage{
data(clone_f_selection_sweep)
}
\description{
Details can be found in the "Time series data" vignette.
}
\keyword{datasets}
