% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_clone_frequency.temulator_result_object}
\alias{get_clone_frequency.temulator_result_object}
\title{Gets clone fractions from a TEMULATOR result object}
\usage{
\method{get_clone_frequency}{temulator_result_object}(x)
}
\arguments{
\item{x}{object of class 'temulator_result_object'.}
}
\value{
a named numeric vector
}
\description{
Gets clone fractions from a TEMULATOR result object
}
