% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mobster_simulations}
\alias{mobster_simulations}
\title{150 synthetic sequencing datasets from MOBSTER}
\format{
A names list of 150 temulator_result_objects.
}
\usage{
data(mobster_simulations)
}
\description{
A list of 150 synthetic sequencing data ('temulator_result_object')
used in Caravagna et al. (2020).
}
\keyword{datasets}
