% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mobster_summary_of_simulations}
\alias{mobster_summary_of_simulations}
\title{Summary statistics of synthetic sequencing datasets used in MOBSTER}
\format{
A data frame with 1755 rows and 8 variables:
\describe{
  \item{subclone_start}{insertion time of the subclone}
  \item{subclone_birthrate}{birth rate of the subclone}
  \item{n}{simulation number (9 per parameter combination)}
  \item{seed}{seed used for the simulation}
  \item{n_mutations_before_insertions}{number of mutations that occured prior to mutation}
  \item{subclone_fraction}{fraction of all cells that subclone made up}
  \item{assigned_id}{id used as identifier in the paper}
}
}
\usage{
data(mobster_summary_of_simulations)
}
\description{
Summary statistics of synthetic sequencing datasets used in MOBSTER
}
\keyword{datasets}
