% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.temulator_result_object}
\alias{plot.temulator_result_object}
\title{Plot method for TEMULATOR result object}
\usage{
\method{plot}{temulator_result_object}(x, quite = FALSE, ...)
}
\arguments{
\item{x}{object of class 'temulator_result_object'.}

\item{quite}{logical indicating if result should be plotted.}

\item{...}{additional parameters passed to get_sequencing_data and assign_mutation_label.}
}
\value{
A ggplot object.
}
\description{
Plot method for TEMULATOR result object
}
