% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.temulator_result_object}
\alias{print.temulator_result_object}
\title{Print method for TEMULATOR result object}
\usage{
\method{print}{temulator_result_object}(x, ...)
}
\arguments{
\item{x}{object of class 'temulator_result_object'.}

\item{...}{unused.}
}
\description{
Print method for TEMULATOR result object
}
