% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_selection_sweep}
\alias{samples_selection_sweep}
\title{Simulated sequencing data of a sweeping clone at various time points}
\format{
A data frame with 14850 rows and 7 variables:
\describe{
  \item{clone}{Cell clone the mutation is part of.}
  \item{alt}{Simulated number of variant reads.}
  \item{depth}{Simulated coverage of mutated site.}
  \item{id}{Mutation identifier.}
  \item{vaf}{Simulated VAF of the mutation.}
  \item{n_reactions}{Total number of reactions at present time.}
  \item{time}{Gillespie time.}
}
}
\usage{
samples_selection_sweep
}
\description{
Details can be found in the "Time series data" vignette.
}
\keyword{datasets}
