% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TEMULATOR_object}
\alias{TEMULATOR_object}
\title{A TEMULATER simulation}
\description{
Type the name of the class to see its methods
}
\section{Fields}{

\describe{
\item{\code{print}}{Print the object}

\item{\code{print_cell_types}}{Print contained cell types}

\item{\code{run}}{Start or restart the simulation \itemize{
\item Parameter: verbose - bool, Flag indicating if progress should be printed.
\item Returns: bool, Indicating success.
}}

\item{\code{sample}}{Sample mutations from the population \itemize{
\item Parameter: min_vaf - double, Minimum VAF reported.
\item Parameter: purity - double, Assumed purity of the tumour.
\item Parameter: depth - double, Sequencing depth.
\item Parameter: depth_model - int,  Distribution of the sequencing depth (1: poisson, 2: overdispersed beta binomial (default), 3: fixed).
\item Returns: data.frame containing mutation information.
}}
}}

