% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_series_data_example}
\alias{time_series_data_example}
\title{Example of a time series dataset for a simulation used in the MOBSTER paper.}
\format{
A data frame with 1100 rows and 5 variables:
\describe{
  \item{reactions}{number of reactions at time point}
  \item{t}{gillespie time at time point}
  \item{cells_a}{number of cells for the ancestral clone  at time point}
  \item{cells_sc}{number of cells for the (selected) subclone  at time point}
  \item{cells}{total number of cells at time point}
}
}
\usage{
time_series_data_example
}
\description{
Example of a time series dataset for a simulation used in the MOBSTER paper.
}
\keyword{datasets}
