% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{validate_temulator_result_object}
\alias{validate_temulator_result_object}
\title{Validates a TEMULATOR result object}
\usage{
validate_temulator_result_object(x)
}
\arguments{
\item{x}{object of class 'temulator_result_object'.}
}
\value{
TRUE if successfull. Throws an error on failure.
}
\description{
Validates a TEMULATOR result object
}
