% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateTumour.R
\name{simulateTumour}
\alias{simulateTumour}
\title{Main function used to produce a simulation}
\usage{
simulateTumour(
  birthrates = c(1, 1),
  deathrates = c(0.2, 0.2),
  mutation_rates = c(16, 16),
  clone_start_times = c(0, 256),
  fathers = c(0, 0),
  simulation_end_time = 1048576,
  seed = 42,
  number_clonal_mutations = 100,
  purity = 1,
  min_vaf = 0.01,
  depth = 100,
  depth_model = 2,
  verbose = FALSE,
  subset_fractions = numeric()
)
}
\arguments{
\item{birthrates}{Birth rate(s) of subpopulations}

\item{deathrates}{Death rate(s) of subpopulations}

\item{mutation_rates}{Mutation rate(s) per division}

\item{clone_start_times}{Timepoint (in number of reaction) a subpopulation is introduced}

\item{fathers}{Subpopulation (0 based) a subpopulations originate from}

\item{simulation_end_time}{Number of reactions performed}

\item{seed}{Simulation seed}

\item{number_clonal_mutations}{Number of clonal variants}

\item{purity}{Assumed purity of the tumour}

\item{min_vaf}{Minimum VAF to report}

\item{depth}{Sequencing depth}

\item{depth_model}{Number specificing the distribution of the sequencing depth to use (0: no sequencing, 1:  poisson, 2: overdispersed beta binomial (default), 3: fixed)}

\item{verbose}{Print progress?}

\item{subset_fractions}{Optional, Numeric vector specificing fraction variants are subset to.}
}
\value{
A temulator_result_object.
}
\description{
Main function used to produce a simulation
}
\examples{
simulateTumour()
simulateTumour(subset_fractions=c("WES"=0.03)) 
simulateTumour(depth=c(30,50,100)) # three different sequencing depths  
}
